# feat: completely rework state tree storage