… CHANGELOG.md
… Cargo.lock
… Cargo.toml
… crates/merkle-tree/src/merkle_node.rs
… crates/merkle-tree/src/transaction.rs
… crates/merkle-tree/src/tree.rs
… crates/pathfinder/Cargo.toml
… crates/pathfinder/src/state/sync.rs
… crates/rpc/Cargo.toml
… crates/rpc/fixtures/mainnet.sqlite
… crates/rpc/src/lib.rs
… crates/rpc/src/pathfinder/methods/get_proof.rs
… crates/storage/Cargo.toml
… crates/storage/src/connection/block.rs
… crates/storage/src/connection/class.rs
… crates/storage/src/connection/state.rs
… crates/storage/src/connection/trie.rs
… crates/storage/src/params.rs
… crates/storage/src/schema.rs
… crates/storage/src/schema/base.rs
… crates/storage/src/schema/revision_0031.rs
… crates/storage/src/schema/revision_0032.rs
… crates/storage/src/schema/revision_0033.rs
… crates/storage/src/schema/revision_0034.rs
… crates/storage/src/schema/revision_0035.rs
… crates/storage/src/schema/revision_0036.rs
… crates/storage/src/schema/revision_0037.rs
… crates/storage/src/schema/revision_0038.rs
… crates/storage/src/schema/revision_0039.rs
… crates/storage/src/schema/revision_0040.rs