This PR improves state tree update performance by entirely reworking how we store trees.

**This is a breaking change and forces a complete database resync**

### What problem is this solving

Our state tree update is running into performance problems, which gets dramatically exaggerated by low IOPS or high latency disks - to the point where these can no longer keep up in particular on mainnet.

Some rough benchmarks:
```
------------------------------
| disk      | before | after |
------------------------------
| local SSD | 3s     | <1s   |
| slow disk | 20-45s | 2-5s  |
------------------------------
```

The current slow-down comes due to the massive size of the tables; and also due to the randomness of the hash-index.

### How?

This PR makes the following (high-level) changes to how we store trees:
- switch from hash index to an integer index for trie node storage
- move leaf storage out of the tree entirely

This has a two-fold impact
1. The tables are *much* smaller
2. Lookup performance is much better due to the less random (and smaller) index

### Database migration

Technically we could attempt to migrate the current database, however this is extremely impractical. Migrating a single block's tree takes hours on a fast disk.

Instead I've chosen to make the database not upgradeable i.e. you have to resync completely.

I've also done some minor schema cleanup while we have the ability to make changes here without regard for backwards compat.

### Knock-on effects

There were quite a few knock-on effects for the merkle-tree implementation itself. I think it came out okay, but the biggest change externally is probably the requirement of a block number for the specific storage, contract and class trees. This comes from externalizing the leaf nodes - this is typically redundant data since we already store all storage updates.

### Pruning

I've ignored pruning entirely here. This performance improvement should be good enough to buy us quite a bit of time.

Pruning can be added by adding reference counting; however tests show that this additional book-keeping does impact performance. I suggest that we have two orthogonal modes:
1. archive - has no pruning logic or book-keeping at all
2. prune   - has pruning logic and book-keeping

The book-keeping performance impact is quite reasonable when the data is kept small i.e. when pruning is actually done. In archive mode the book-keeping cost would be significant and unused -- hence my suggestion to separate them entirely.

This will be tackled in a different PR. This PR only has archive mode -- the same as we currently have.

### Next steps

I want to re-run this on all the networks to ensure nothing has changed since we first ran the tests.

Then we should create snapshots for every network and then release.

And also re-add the parallel contract tree update processing that @kkovaacs added on his branch.