@Mirko-von-Leipzig 2023-10-04
Ah and the index sizes must be larger than u32. I'll change that and probably make them newtypes as @pierre-l suggested.