@Mirko-von-Leipzig 2023-10-04
> Few (likely dumb) questions:
> 
> Do I understand correctly, the main reason of the database size blowing up is the storage of merkle-tree nodes for contracts classes and state, right?

Yes currently roughly half the db is just storing these nodes. 

> Just wondering if offloading this from SQLite to persistent tree storage based on say rocksdb will make it way more performant and reduce the IO pressure on SQLite significantly? Rocksdb et al should be able to make merkle tree calculations mostly CPU-bound, as it traverses sorted data extremely fast with moderate IO pressure.

Its possible but that's essentially what this PR is doing - making the data sorted instead of random. The table change is effectively this:

```
before
---------------------------
| hash (PK) | data (BLOB) |
---------------------------
after
--------------------------------
| idx (PK) | hash | data (BLOB)|
--------------------------------
```
Where the data encoding can now change to something approximately like:
```rust
enum Before {
  BinaryNode { left: Felt, right: Felt },
  EdgeNode { child: Felt, path: Felt },
}

enum After {
  BinaryNode { left: u64, right: u64},
  EdgeNode { child: u64, path: Felt },
}
```
The data sizes are much smaller afterwards; usually in the 3B to 35B range using varint encoding. Whereas before the minimum was about 100B per node.

There is potentially some value in a different db just for this; but it will come with additional work since you lose having a single transaction.

> Also just wondering if it makes sense to force the user to choose an appropriate security/performance tradeoff ratio by making full validation (hashes, commitments) configurable? Might get very tricky once P2P kicks off properly.

I think there are certain tradeoffs one can make there - eventually. Once p2p arrives properly all these options change substantially. Right now if you want insecure data you may as well just use an API service.